[feature] 팩토리에서 좌표와 아이디를 생성하여 사각형을 만들도록 함
